simple model for topcoder

+++ datasets/torch_movielen.py
@@ -187,7 +187,7 @@ class TorchMovielen10k(DataBunch):
         feat_index_list = feat_matrix.nonzero()
         feat_index_array: np.ndarray = np.vstack(feat_index_list)
 
-        feat_index = T.tensor(feat_index_array.tolist())
+        feat_index = T.tensor(feat_index_array.tolist(), dtype=T.long)
         feat_value = T.tensor(feat_matrix.data, dtype=T.double)
         feat_tensor = T.sparse_coo_tensor(feat_index_list,
                                           feat_value,
